tools/gomoku/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
github3.py==2.0.0
idna==2.10
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
